refactor(learning): split admin dashboard into single-responsibility components
- Extract EmployeeRow, CompletionFunnel, CategoryPerformance, and
  TopTrainings into dedicated files (one component per file)
- Move shared helpers, status maps, and avatar colors to admin-constants
- Slim admin-dashboard orchestrator from 821 to ~230 lines
- Update barrel exports in components/index.ts
